ingestion/requirements.txt
asgiref==3.2.3
asn1crypto==1.3.0
attrs==19.3.0
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
Django==3.0.2
django-taggit==1.2.0
djangorestframework==3.11.0
djangorestframework-api-key==1.4.1
fuzzywuzzy==0.17.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.2
Markdown==3.1.1
musicbrainzngs==0.7.1
packaging==20.1
parsel==1.5.2
Protego==0.1.16
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pylast==3.2.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
queuelib==1.5.0
requests==2.22.0
Scrapy==1.8.0
scrapy-djangoitem==1.1.1
service-identity==18.1.0
six==1.14.0
spotipy==2.7.1
sqlparse==0.3.0
Twisted==19.10.0
urllib3==1.25.8
w3lib==1.21.0
webencodings==0.5.1
zope.interface==4.7.1
